Create new checksum module.
